requirements.txt
cairocffi==1.7.1
CairoSVG==2.8.2
cffi==1.17.1
chardet==5.2.0
contourpy==1.3.2
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
easyocr==1.7.2
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.0
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyclipper==1.3.0.post6
pycparser==2.22
PyMuPDF==1.25.5
pyparsing==3.2.3
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
shapely==2.1.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.4.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
webencodings==0.5.1
